feat: refactor payment services and enhance error handling with shared utilities
