+++ src/utils/index.ts
@@ -4,6 +4,5 @@ export * from './transform.utils'
 export * from './jwt.utils'
 export * from './email.utils'
 export * from './file.utils'
-export * from './persistence.utils'
-export * from './ngrok'
 export * from './redis.utils'
+export * from './ngrok'
